Add module and process subsystems.
As the process subsystem needs to make hypercalls the hypercall initializtion
has moved into DriverEntry(). This is ok as it will only ever be called if
the xen driver is not going to conflict with other PV drivers (otherwise
installation would have been vetoed by the co-installer before the service
was added).
Other stuff was moved fixed and changed as I went along; perticularely the
PDO naming code seemed to be wrong.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>